+++ swing/framework-ui/src/test/java/is/codion/swing/framework/ui/EntityTableCellRendererFactoryTest.java
@@ -21,7 +21,9 @@ package is.codion.swing.framework.ui;
 import is.codion.common.user.User;
 import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.db.local.LocalEntityConnectionProvider;
+import is.codion.framework.domain.entity.attribute.Attribute;
 import is.codion.swing.common.ui.component.table.FilterTableCellRenderer;
+import is.codion.swing.common.ui.component.table.FilterTableColumn;
 import is.codion.swing.framework.model.SwingEntityTableModel;
 import is.codion.swing.framework.ui.TestDomain.Department;
 import is.codion.swing.framework.ui.TestDomain.Employee;
@@ -30,7 +32,7 @@ import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-public class EntityTableCellRendererTest {
+public class EntityTableCellRendererFactoryTest {
 
 	private static final User UNIT_TEST_USER =
 					User.parse(System.getProperty("codion.test.user", "scott:tiger"));
@@ -44,7 +46,9 @@ public class EntityTableCellRendererTest {
 	void test() {
 		EntityTablePanel tablePanel = new EntityTablePanel(new SwingEntityTableModel(Employee.TYPE, CONNECTION_PROVIDER));
 		tablePanel.tableModel().refresh();
-		FilterTableCellRenderer renderer = EntityTableCellRenderer.builder(tablePanel.tableModel(), Employee.NAME).build();
+		EntityTableCellRendererFactory factory = new EntityTableCellRendererFactory(tablePanel.tableModel());
+		FilterTableColumn<Attribute<?>> column = tablePanel.table().columnModel().column(Employee.NAME);
+		FilterTableCellRenderer renderer =  factory.tableCellRenderer(column);
 		renderer.getTableCellRendererComponent(tablePanel.table(), null, false, false, 0, 0);
 		renderer.getTableCellRendererComponent(tablePanel.table(), null, true, false, 0, 0);
 		renderer.getTableCellRendererComponent(tablePanel.table(), null, true, true, 0, 0);
@@ -60,8 +64,10 @@ public class EntityTableCellRendererTest {
 
 	@Test
 	void entityMismatch() {
-		EntityTablePanel tablePanel = new EntityTablePanel(new SwingEntityTableModel(Employee.TYPE, CONNECTION_PROVIDER));
-		tablePanel.tableModel().refresh();
-		assertThrows(IllegalArgumentException.class, () -> EntityTableCellRenderer.builder(tablePanel.tableModel(), Department.NAME));
+		EntityTablePanel tablePanel = new EntityTablePanel(new SwingEntityTableModel(Department.TYPE, CONNECTION_PROVIDER));
+		FilterTableColumn<Attribute<?>> column = tablePanel.table().columnModel().column(Department.NAME);
+
+		EntityTableCellRendererFactory factory = new EntityTableCellRendererFactory(new SwingEntityTableModel(Employee.TYPE, CONNECTION_PROVIDER));
+		assertThrows(IllegalArgumentException.class, () -> factory.tableCellRenderer(column));
 	}
 }